+++ instrumentation_test.go
@@ -15,32 +15,41 @@
 package auto
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 )
 
 func TestWithServiceName(t *testing.T) {
+	ctx := context.Background()
 	testServiceName := "test_serviceName"
 
 	// Use WithServiceName to config the service name
-	c := newInstConfig([]InstrumentationOption{WithServiceName((testServiceName))})
+	c, err := newInstConfig(ctx, []InstrumentationOption{WithServiceName((testServiceName))})
+	require.NoError(t, err)
 	assert.Equal(t, testServiceName, c.serviceName)
 
 	// No service name provided - check for default value
-	c = newInstConfig([]InstrumentationOption{})
+	c, err = newInstConfig(ctx, []InstrumentationOption{})
+	require.NoError(t, err)
 	assert.Equal(t, c.defualtServiceName(), c.serviceName)
 }
 
 func TestWithPID(t *testing.T) {
-	c := newInstConfig([]InstrumentationOption{WithPID(1)})
+	ctx := context.Background()
+
+	c, err := newInstConfig(ctx, []InstrumentationOption{WithPID(1)})
+	require.NoError(t, err)
 	assert.Equal(t, 1, c.target.Pid)
 
 	const exe = "./test/path/program/run.go"
 	// PID should override valid target exe
-	c = newInstConfig([]InstrumentationOption{WithTarget(exe), WithPID(1)})
+	c, err = newInstConfig(ctx, []InstrumentationOption{WithTarget(exe), WithPID(1)})
+	require.NoError(t, err)
 	assert.Equal(t, 1, c.target.Pid)
 	assert.Equal(t, "", c.target.ExePath)
 }
@@ -49,7 +58,8 @@ func TestWithEnv(t *testing.T) {
 	t.Run("OTEL_GO_AUTO_TARGET_EXE", func(t *testing.T) {
 		const path = "./test/path/program/run.go"
 		mockEnv(t, map[string]string{"OTEL_GO_AUTO_TARGET_EXE": path})
-		c := newInstConfig([]InstrumentationOption{WithEnv()})
+		c, err := newInstConfig(context.Background(), []InstrumentationOption{WithEnv()})
+		require.NoError(t, err)
 		assert.Equal(t, path, c.target.ExePath)
 		assert.Equal(t, 0, c.target.Pid)
 	})
@@ -57,7 +67,8 @@ func TestWithEnv(t *testing.T) {
 	t.Run("OTEL_SERVICE_NAME", func(t *testing.T) {
 		const name = "test_service"
 		mockEnv(t, map[string]string{"OTEL_SERVICE_NAME": name})
-		c := newInstConfig([]InstrumentationOption{WithEnv()})
+		c, err := newInstConfig(context.Background(), []InstrumentationOption{WithEnv()})
+		require.NoError(t, err)
 		assert.Equal(t, name, c.serviceName)
 	})
 
@@ -65,7 +76,8 @@ func TestWithEnv(t *testing.T) {
 		const name = "test_service"
 		val := fmt.Sprintf("a=b,fubar,%s=%s,foo=bar", semconv.ServiceNameKey, name)
 		mockEnv(t, map[string]string{"OTEL_RESOURCE_ATTRIBUTES": val})
-		c := newInstConfig([]InstrumentationOption{WithEnv()})
+		c, err := newInstConfig(context.Background(), []InstrumentationOption{WithEnv()})
+		require.NoError(t, err)
 		assert.Equal(t, name, c.serviceName)
 	})
 }
@@ -88,7 +100,8 @@ func TestOptionPrecedence(t *testing.T) {
 			WithServiceName("wrong"),
 			WithEnv(),
 		}
-		c := newInstConfig(opts)
+		c, err := newInstConfig(context.Background(), opts)
+		require.NoError(t, err)
 		assert.Equal(t, path, c.target.ExePath)
 		assert.Equal(t, 0, c.target.Pid)
 		assert.Equal(t, name, c.serviceName)
@@ -106,7 +119,8 @@ func TestOptionPrecedence(t *testing.T) {
 			WithPID(1),
 			WithServiceName(name),
 		}
-		c := newInstConfig(opts)
+		c, err := newInstConfig(context.Background(), opts)
+		require.NoError(t, err)
 		assert.Equal(t, "", c.target.ExePath)
 		assert.Equal(t, 1, c.target.Pid)
 		assert.Equal(t, name, c.serviceName)